+++ R/plots.R
@@ -19,33 +19,43 @@
 #'   will reorder the groups in such a way that the majority group actually comes first.
 #'   If you want to control the ordering yourself, use "majority_fixed" and specify
 #'   the \code{group} variable as a factor variable.
+#' @param secondary_plot If \code{NULL} (default), no secondary plot is drawn.
+#'   If "segregation", a secondary plot is drawn that shows adjusted local segregation
+#'   scores for each unit. If "cumulative", a secondary plot is drawn that shows
+#'   the cumulative contribution of each unit toward the total H (calculated as the
+#'   proportion of each unit times the adjusted local segregation of each unit)0.
 #' @param reference_distribution Specifies the reference distribution, given as
 #'   a two-column data frame, to be plotted on the right.
 #'   If order is \code{segregation}, then this reference distribution is
 #'   also used to compute the local segregation scores.
 #' @param bar_space Specifies space between single units.
-#' @param title Adds a plot title and appends the value of the H index.
-#' @return Returns a ggplot2 object.
+#' @return Returns a ggplot2 or patchwork object.
 #' @import data.table
 #' @export
-segplot <- function(data, group, unit, weight, order = "segregation",
+segplot <- function(data, group, unit, weight,
+                    order = "segregation", secondary_plot = NULL,
                     reference_distribution = NULL,
-                    bar_space = 0, title = NULL) {
+                    bar_space = 0) {
+    if (!requireNamespace("patchwork", quietly = TRUE)) {
+        stop("Please install patchwork to use this function")
+    }
     if (!requireNamespace("ggplot2", quietly = TRUE)) {
         stop("Please install ggplot2 to use this function")
     }
 
-    stopifnot(length(group) == 1)
-    stopifnot(length(unit) == 1)
-    stopifnot(order %in% c("segregation", "entropy", "majority", "majority_fixed"))
-
+    checkmate::assert_character(group, len = 1)
+    checkmate::assert_character(unit, len = 1)
     d <- prepare_data(data, group, unit, weight)
     # easier if renamed
     setnames(d, group, "group")
     setnames(d, unit, "unit")
 
-    d[, unit := as.character(unit)]
+    # check other arguments
+    checkmate::assert_choice(order, c("segregation", "entropy", "majority", "majority_fixed"))
+    checkmate::assert_choice(secondary_plot, c("segregation", "cumulative"), null.ok = TRUE)
+    checkmate::assert_data_frame(reference_distribution, ncols = 2, nrows = d[, uniqueN(group)], null.ok = TRUE)
 
+    d[, unit := as.character(unit)]
     d[, p := freq / sum(freq), by = .(unit)]
     d[, p_unit := sum(freq), by = .(unit)]
     N <- d[, first(p_unit), by = .(unit)][, sum(V1)]
@@ -56,10 +66,7 @@ segplot <- function(data, group, unit, weight, order = "segregation",
         overall <- d[, .(freq = sum(freq)), by = .(group)]
         overall[, p := freq / sum(freq)]
     } else {
-        stopifnot(is.data.frame(reference_distribution))
-        stopifnot(ncol(reference_distribution) == 2)
         stopifnot(names(reference_distribution) == c(group, "p"))
-        stopifnot(nrow(reference_distribution) == d[, uniqueN(group)])
         overall <- as.data.table(reference_distribution)
         setnames(overall, group, "group")
     }
@@ -141,13 +148,50 @@ segplot <- function(data, group, unit, weight, order = "segregation",
         ) +
         ggplot2::labs(fill = NULL)
 
-    if (order == "segregation") {
+    if (order == "segregation" && is.null(secondary_plot)) {
         plot <- plot + ggplot2::labs(x = "< more segregated | less segregated >")
-    }
+    } else if (order == "segregation") {
+        entropy <- entropy(overall, "group", weight = "p")
+        H_index <- wide[, sum(ls * p_unit) / entropy]
 
-    if (!is.null(title)) {
-        H <- mutual_total(d, "group", "unit", weight = "freq")[stat == "H", est]
-        plot <- plot + ggplot2::labs(title = paste0(title, " (H = ", round(H, 2), ")"))
+        if (secondary_plot == "segregation") {
+            wide[, stat := ls / entropy]
+            label <- "adj. LS"
+        } else if (secondary_plot == "cumulative") {
+            wide[, stat := cumsum(p_unit * ls / entropy)]
+            wide[, stat := max(stat) - stat + min(stat)]
+            label <- "Cumulative"
+        }
+
+        stat_segments_h <- wide[, .(unit, x = xmin, xend = xmax, y = stat, yend = stat)]
+        stat_segments_v <- wide[, .(unit, x = xmax, xend = xmax, y = stat, yend = shift(stat, type = "lead"))]
+        stat_segments <- rbindlist(list(stat_segments_h, stat_segments_v))
+        sub_plot <- ggplot2::ggplot(
+            mapping = ggplot2::aes(x = x, y = y, xend = xend, yend = yend)
+        ) +
+            ggplot2::geom_vline(xintercept = wide[, max(xmax)], linewidth = 0.2) +
+            ggplot2::geom_hline(yintercept = H_index, color = "orange") +
+            ggplot2::geom_segment(data = stats::na.omit(stat_segments)) +
+            ggplot2::theme_bw() +
+            ggplot2::scale_x_continuous(
+                limits = c(0, overall[, max(xmax)]), expand = c(0, 0),
+                labels = function(x) scales::label_percent(1)(1 - x)
+            ) +
+            ggplot2::scale_y_continuous(sec.axis = ggplot2::dup_axis(
+                breaks = H_index,
+                labels = function(x) paste0("H = ", round(H_index, 3))
+            )) +
+            ggplot2::theme(
+                panel.grid.major.x = ggplot2::element_blank(),
+                panel.grid.minor.x = ggplot2::element_blank(),
+                panel.grid.minor.y = ggplot2::element_blank(),
+                axis.text.x = ggplot2::element_blank(),
+                axis.ticks.x = ggplot2::element_blank(),
+                axis.title.y.right = ggplot2::element_blank()
+            ) +
+            ggplot2::labs(y = label, x = "< more segregated | less segregated >")
+
+        plot <- plot / sub_plot + patchwork::plot_layout(heights = c(4, 1))
     }
 
     plot